api/requirements.txt
# Core FastAPI Framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Data Validation & Serialization
pydantic==2.11.7
pydantic_core==2.33.2
email-validator==2.2.0

# HTTP Client & Server (let pip resolve compatible versions)
httpx>=0.26,<0.28
httpcore>=1.0.0
h11>=0.14.0
anyio>=4.0.0

# Database & External Services
supabase==2.10.0

# Environment & Configuration
python-dotenv==1.1.1

# Type System & Annotations
typing-extensions>=4.0.0
typing-inspection>=0.4.0
annotated-types>=0.6.0

# Utilities
click>=8.0.0
certifi>=2024.0.0
idna>=3.0
sniffio>=1.3.0